refactor: extract agent-link hooks into agent-link-hooks.ts (commit 9)
