+++ chatbot/lib/agent-link-hooks.ts
@@ -1,0 +1,250 @@
+/**
+ * Agent-link-specific hook matchers for the sub-agent query().
+ *
+ * Extracted from subagent-hooks.ts so the justification-gate logic
+ * (chat_to / review_with / escalate_to reflection, group silence, handoff
+ * consideration) can be omitted by embedders that don't use agent links.
+ */
+
+import type {
+  PreToolUseHookSpecificOutput,
+  PostToolUseHookSpecificOutput,
+  HookCallbackMatcher,
+} from "@anthropic-ai/claude-agent-sdk";
+import { ESCALATE_PATTERNS, HANDOFF_PATTERNS, REVIEW_PATTERNS } from "@/lib/agent-link-patterns";
+import {
+  CONFIDENCE_THRESHOLD,
+  impactPlaceholder,
+  startChatToToolName,
+  startReviewWithToolName,
+  startEscalateToToolName,
+  awaitChatToToolName,
+  awaitReviewWithToolName,
+  awaitEscalateToToolName,
+  awaitRunScriptToolName,
+} from "@/lib/agent-tools";
+import { getMcpStructured } from "@/lib/hooks";
+
+// ─── Reflection gate ──────────────────────────────────────────────────────────
+
+const thresholdTable = Object.entries(CONFIDENCE_THRESHOLD)
+  .map(
+    ([k, { threshold, description }]) =>
+      `  ${k.padEnd(10)}→ ${threshold === Infinity ? "never handed off" : `confidence ≥ ${threshold}`}  (${description})`,
+  )
+  .join("\n");
+
+function buildReflectionPrompt(patterns: string, isGroupMode?: boolean): string {
+  const ifNo = isGroupMode
+    ? "If NO to any: do not re-call. Continue without the handoff and DO NOT explain your reasoning."
+    : "If NO to any: do not re-call. Respond directly with your results.";
+  return `<reminder>
+You are about to delegate work to another agent. Pause and answer:
+
+1. Have you completed your own task? If not, finish it first.
+2. Do you have concrete, specific output to hand off (issues list, report, IDs, findings)?
+   Vague intent ("check this", "take a look") is not a valid handoff.
+3. Does the handoff match a clear pattern?
+<patterns>
+${patterns}
+</patterns>
+
+If YES to all: re-call this tool with a \`justification\` object:
+  {
+    "impact": "<${impactPlaceholder}>",
+    "pattern": "<which pattern above applies>",
+    "handoff": "<one sentence: what concrete output you are handing off>",
+    "confidence": <0–1>
+  }
+${isGroupMode ? "Do NOT output and respond with any text such as narration, status updates, or confirmations.\n" : ""}
+Impact-gated thresholds:
+${thresholdTable}
+
+${ifNo}
+</reminder>`;
+}
+
+function makeReflectionMatcher(matcher: string, reflectionPrompt: string): HookCallbackMatcher {
+  return {
+    matcher,
+    hooks: [
+      async (input) => {
+        if (input.hook_event_name !== "PreToolUse") return { continue: true };
+        const { tool_input } = input;
+        const inputObj = tool_input !== null && typeof tool_input === "object" ? tool_input : null;
+
+        const justificationRaw: unknown = inputObj
+          ? Reflect.get(inputObj, "justification")
+          : undefined;
+        if (
+          justificationRaw === null ||
+          justificationRaw === undefined ||
+          typeof justificationRaw !== "object"
+        ) {
+          const hookSpecificOutput: PreToolUseHookSpecificOutput = {
+            hookEventName: "PreToolUse",
+            permissionDecision: "deny",
+            permissionDecisionReason: reflectionPrompt,
+          };
+          return { hookSpecificOutput };
+        }
+
+        const confidence: unknown = Reflect.get(justificationRaw, "confidence");
+        const impact: unknown = Reflect.get(justificationRaw, "impact");
+
+        const impactKey =
+          typeof impact === "string" && impact in CONFIDENCE_THRESHOLD ? impact : undefined;
+
+        if (impactKey === undefined) {
+          const hookSpecificOutput: PreToolUseHookSpecificOutput = {
+            hookEventName: "PreToolUse",
+            permissionDecision: "deny",
+            permissionDecisionReason: `impact is missing or invalid ("${String(impact)}"). Re-call with impact: ${impactPlaceholder}.`,
+          };
+          return { hookSpecificOutput };
+        }
+
+        const entry = CONFIDENCE_THRESHOLD[impactKey];
+
+        if (entry.threshold === Infinity) {
+          const hookSpecificOutput: PreToolUseHookSpecificOutput = {
+            hookEventName: "PreToolUse",
+            permissionDecision: "deny",
+            permissionDecisionReason: `Low-impact handoffs are skipped. Handle via message instead, or raise the impact level if this is genuinely consequential.`,
+          };
+          return { hookSpecificOutput };
+        }
+
+        if (typeof confidence !== "number" || confidence < entry.threshold) {
+          const hookSpecificOutput: PreToolUseHookSpecificOutput = {
+            hookEventName: "PreToolUse",
+            permissionDecision: "deny",
+            permissionDecisionReason:
+              typeof confidence !== "number"
+                ? `confidence is missing or not a number. Re-call with a numeric confidence score (0–1).`
+                : `Confidence ${confidence} is below the required threshold of ${entry.threshold} for ${impactKey} impact. Only proceed if genuinely confident this handoff is necessary and well-scoped.`,
+          };
+          return { hookSpecificOutput };
+        }
+
+        const hookSpecificOutput: PreToolUseHookSpecificOutput = {
+          hookEventName: "PreToolUse",
+          permissionDecision: "allow",
+        };
+        return { hookSpecificOutput };
+      },
+    ],
+  };
+}
+
+export function buildReflectionMatchers(isGroupMode?: boolean): HookCallbackMatcher[] {
+  return [
+    makeReflectionMatcher(
+      `mcp__agents__${startChatToToolName(".*")}`,
+      buildReflectionPrompt(HANDOFF_PATTERNS(), isGroupMode),
+    ),
+    makeReflectionMatcher(
+      `mcp__agents__${startReviewWithToolName(".*")}`,
+      buildReflectionPrompt(REVIEW_PATTERNS(), isGroupMode),
+    ),
+    makeReflectionMatcher(
+      `mcp__agents__${startEscalateToToolName(".*")}`,
+      buildReflectionPrompt(ESCALATE_PATTERNS(), isGroupMode),
+    ),
+  ];
+}
+
+// ─── Stop hook: handoff consideration ────────────────────────────────────────
+
+export function buildHandoffConsiderationPrompt(
+  tools: Array<{ name: string; description: string }>,
+  isGroupMode?: boolean,
+  lastAssistantMessage?: string,
+): string {
+  const toolsXml = tools
+    .map(
+      (t) =>
+        `  <tool>\n    <name>${t.name}</name>\n    <description>${t.description}</description>\n  </tool>`,
+    )
+    .join("\n");
+  const ifNo = isGroupMode
+    ? "If no: stop immediately and DO NOT explain your reasoning."
+    : `If no: respond with exactly:\n"${lastAssistantMessage ?? ""}"`;
+  return `<reminder>
+Before you finish: have you considered whether to hand off your results to a linked agent?
+
+<handoff_tools>
+${toolsXml}
+</handoff_tools>
+
+If yes: call the appropriate tool before stopping.
+${isGroupMode ? "Do NOT output and respond with any text such as narration, status updates, or confirmations.\n" : ""}${ifNo}
+</reminder>`;
+}
+
+// ─── Group chat silence hooks ─────────────────────────────────────────────────
+
+const GROUP_START_MATCHER = [
+  `mcp__agents__${startChatToToolName(".*")}`,
+  `mcp__agents__${startReviewWithToolName(".*")}`,
+  `mcp__agents__${startEscalateToToolName(".*")}`,
+].join("|");
+
+const GROUP_AWAIT_MATCHER = [
+  `mcp__agents__${awaitChatToToolName(".*")}`,
+  `mcp__agents__${awaitReviewWithToolName(".*")}`,
+  `mcp__agents__${awaitEscalateToToolName(".*")}`,
+].join("|");
+
+const GROUP_START_SILENCE = `<reminder>
+You have started a handoff task.
+Call the corresponding await tool immediately to collect the result.
+Do NOT output and respond with any text such as narration, status updates, or confirmations before calling the await tool.
+</reminder>`;
+
+const GROUP_AWAIT_SILENCE = `<reminder>
+Do NOT output any text or narration between tool calls.
+Proceed directly to the next tool call.
+Only deliver your response when you have fully completed your task with concrete results to share.
+</reminder>`;
+
+function makeGroupSilenceHook(matcher: string, context: string): HookCallbackMatcher {
+  return {
+    matcher,
+    hooks: [
+      async (input) => {
+        if (input.hook_event_name !== "PostToolUse") return { continue: true };
+        const hookSpecificOutput: PostToolUseHookSpecificOutput = {
+          hookEventName: "PostToolUse",
+          additionalContext: context,
+        };
+        return { hookSpecificOutput };
+      },
+    ],
+  };
+}
+
+export const groupStartHandoffHook = makeGroupSilenceHook(GROUP_START_MATCHER, GROUP_START_SILENCE);
+export const groupAwaitHandoffHook = makeGroupSilenceHook(GROUP_AWAIT_MATCHER, GROUP_AWAIT_SILENCE);
+
+export function makeGroupScriptAwaitToneHook(manifestKey: string): HookCallbackMatcher {
+  return {
+    matcher: `mcp__agents__${awaitRunScriptToolName(manifestKey)}`,
+    hooks: [
+      async (input) => {
+        if (input.hook_event_name !== "PostToolUse") return { continue: true };
+        const structured = getMcpStructured(input.tool_response);
+        const status =
+          typeof structured === "object" && structured !== null && "status" in structured
+            ? (structured as { status: unknown }).status
+            : undefined;
+        if (status === "still_running") return { continue: true };
+        const hookSpecificOutput: PostToolUseHookSpecificOutput = {
+          hookEventName: "PostToolUse",
+          additionalContext: `<reminder>Respond in your own voice and tone as defined by your agent script role.</reminder>`,
+        };
+        return { hookSpecificOutput };
+      },
+    ],
+  };
+}
